增加日志插入功能

+++ st-core/src/main/java/com/wtf/core/domain/model/IntegralLog.java
@@ -10,6 +10,6 @@ import lombok.ToString;
 @ToString
 public class IntegralLog extends BaseModel {
 
-    private double count;
-    private int userId;
+    private Double count;
+    private Long userId;
 }
